+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to
   payments of installments that are in the past
 - Deprecated field `has_consent_to_terms` for `Order` model
 
+### Fixed
+
+- Prevent duplicate Address objects for a user or an organization
+
 ### Removed
 
 - Remove the `has_consent_to_terms` field from the `Order` edit view